+++ frontend/src/components/Page.tsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Page = styled.div`
-  max-width: 800px;
-  margin: auto;
-  padding: 2em;
-`;