+++ frontend/src/components/molecules/FloatingResetButton.tsx
@@ -1,12 +1,13 @@
 import { Button } from "@blueprintjs/core";
 import * as React from "react";
-import FloatingCard from "./FloatingCard";
+import { FloatingCard } from "../atoms/";
 
 interface IFloatingResetButtonProps {
   onClick?: () => any;
 }
-export const FloatingResetButton: React.FC<IFloatingResetButtonProps> = ({ onClick }) => (
+const FloatingResetButton: React.FC<IFloatingResetButtonProps> = ({ onClick }) => (
   <FloatingCard>
     <Button icon="compass" onClick={onClick} />
   </FloatingCard>
 );
+export default FloatingResetButton;