+++ frontend/src/components/organisms/index.ts
@@ -1,0 +1,4 @@
+export { default as Graph } from "./Graph";
+export { default as Sidebar } from "./Sidebar";
+export { default as Nav } from "./Nav";
+export { default as InstanceSearch } from "./InstanceSearch";