+++ frontend/src/constants.tsx
@@ -3,3 +3,8 @@ export const DESKTOP_WIDTH_THRESHOLD = 800;
 
 export const DEFAULT_NODE_COLOR = "#CED9E0";
 export const SELECTED_NODE_COLOR = "#48AFF0";
+
+export const INSTANCE_DOMAIN_PATH = "/instance/:domain";
+export interface IInstanceDomainPath {
+  domain: string;
+}