+++ frontend/src/util.ts
@@ -1,4 +1,7 @@
+import { createMatchSelector } from "connected-react-router";
 import fetch from "cross-fetch";
+import { IInstanceDomainPath, INSTANCE_DOMAIN_PATH } from "./constants";
+import { IAppState } from "./redux/types";
 
 let API_ROOT = "http://localhost:4000/api/";
 if (["true", true, 1, "1"].indexOf(process.env.REACT_APP_STAGING || "") > -1) {
@@ -12,3 +15,5 @@ export const getFromApi = (path: string): Promise<any> => {
   path = path.endsWith("/") ? path : path + "/";
   return fetch(domain + path).then(response => response.json());
 };
+
+export const domainMatchSelector = createMatchSelector<IAppState, IInstanceDomainPath>(INSTANCE_DOMAIN_PATH);